Rename to PinterestApi
- PinterestClient sẽ bao gồm các action của Pinterest.
  + Mỗi action là 1 tổ hợp các lệnh của PinterestClient
  + VD: findAnUser = 1 vài lệnh autocomplete + search + ...
  + VD: repin = getPinDetail + getBoardsOfMe + repin
- PinterestApi (trước đây là PinterestClient) sẽ bao gồm các HTTP
request gửi lên server.
- Bot sẽ sử dụng các action ở trong PinterestClient, cái này sẽ được
move trong PR sau.
